requirements.txt
attrs==19.3.0
certifi==2019.9.11
importlib-metadata==0.23
lxml==4.4.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
MolVS==0.1.1
more-itertools==7.2.0
olefile==0.46
packaging==19.2
Pillow==6.2.1
pluggy==0.13.1
progressbar==2.5
progressbar2==3.47.0
py==1.8.0
pyparsing==2.4.5
pytest==5.3.0
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2019.3
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
wcwidth==0.1.7
zipp==0.6.0
